% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare.matrix}
\alias{compare.matrix}
\title{compare.matrix}
\usage{
compare.matrix(input.matrix, output.matrix,
  description = "Number of removed features:")
}
\arguments{
\item{input.matrix}{A numeric matrix as input.}

\item{output.matrix}{A numeric matrix as input.}

\item{description}{A brief description to print out.}
}
\value{
A message object.
}
\description{
compare.matrix
}
