% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAAindex.R
\name{computeAAindexValueByAccession}
\alias{computeAAindexValueByAccession}
\title{computeAAindexValueByAccession}
\usage{
computeAAindexValueByAccession(seq, accession)
}
\arguments{
\item{seq}{Amino acid sequence}

\item{accession}{Valid AAindex accession code}
}
\description{
This function compute an amino acid sequence AAindex property given an accession code.
}

