% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAAindex.R
\name{computeAllAAindexValues}
\alias{computeAllAAindexValues}
\title{computeAllAAindexValues}
\usage{
computeAllAAindexValues(seq)
}
\arguments{
\item{seq}{Amino acid sequence}
}
\description{
This function compute all AAindex values given an amino acid sequence.
}

