% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_functions.R
\name{filter.corr}
\alias{filter.corr}
\title{filter.corr}
\usage{
filter.corr(features, class, mincorr = 0.3)
}
\arguments{
\item{features}{A numeric matrix as input.}

\item{class}{Response variable as numeric vector.}

\item{mincorr}{Correlation coeficient cutoff used for filtering.}
}
\value{
A matrix filtered.
}
\description{
Filter features based on its correletion with a predictor variable.
}
\examples{
output <- filter.corr(features, class, .30)

}
