% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_functions.R
\name{filter.gain.inf}
\alias{filter.gain.inf}
\title{filter.gain.inf}
\usage{
filter.gain.inf(features, class, n.percent = 0.75, zero.gain.out = TRUE)
}
\arguments{
\item{features}{A numeric matrix as input.}

\item{class}{Response variable as numeric vector.}

\item{n.percent}{The percent of features with more gain to be returned.}

\item{zero.gain.out}{Is TRUE (default), zero-gain features will be filtered out (n.percent will be ignored).}
}
\value{
A matrix filtered.
}
\description{
Gain information filter implementation.
}
