% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_functions.R
\name{filter.matrix.corr}
\alias{filter.matrix.corr}
\title{filter.matrix.corr}
\usage{
filter.matrix.corr(features, maxcorr = 0.75)
}
\arguments{
\item{features}{A numeric matrix as input.}

\item{maxcorr}{Correlation coeficient cutoff used for filtering.}
}
\value{
A matrix filtered.
}
\description{
Compute matrix correlation between features and filter using a threshold.
}
\examples{
output <- filter.matrix.corr(features, maxcorr = 0.75)

}
