% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_functions.R
\name{filter.pca}
\alias{filter.pca}
\title{filter.pca}
\usage{
filter.pca(features, center = TRUE, scale = TRUE, cum.var.cutoff = 1)
}
\arguments{
\item{features}{A numeric matrix as input.}

\item{center}{Must the features be centered (default TRUE).}

\item{scale}{Must the features be scaled (default TRUE).}

\item{cum.var.cutoff}{Cumulative variance proportion cutoff. Used to return the Principal Componets (PCs) which satisfies that value. If 1 (default), all PCs will be returned.}
}
\value{
A matrix of Principal components.
}
\description{
Compute Principal Components from input matrix. Control how many components must be returned.
}
