% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterMissingnessRate}
\alias{filterMissingnessRate}
\title{filterMissingnessRate}
\usage{
filterMissingnessRate(mx, max_missing_rate = 0.25)
}
\arguments{
\item{mx}{A numeric matrix as input.}

\item{max_missing_rate}{Maximal missing rate allow for a feature (default: 0.25).}
}
\value{
Filtered matrix with features (col) passing the max missing rate threshold removed.
}
\description{
filterMissingnessRate
}
