% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAAindex.R
\name{getAAindexMatrix}
\alias{getAAindexMatrix}
\title{getAAindexMatrix}
\usage{
getAAindexMatrix()
}
\description{
This function reduce AAindex DataBase and get a simplified features(descriptors) matrix.
}

