% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAAindex.R
\name{getAAindexTable}
\alias{getAAindexTable}
\title{getAAindexTable}
\usage{
getAAindexTable(accession)
}
\arguments{
\item{accession}{Valid AAindex accession code}
}
\description{
Retrieving basic information from AAindex database given an accession code
}

