% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMissingnessRateByFeatures}
\alias{getMissingnessRateByFeatures}
\title{getMissingnessRateByFeatures}
\usage{
getMissingnessRateByFeatures(mx)
}
\arguments{
\item{mx}{A numeric matrix as input.}
}
\value{
A dataframe with features and missingness rate.
}
\description{
Given an input matrix, compute the missingness rate for every feature (expected in cols)
}
