% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE48760}
\alias{GSE48760}
\title{GSE48760 dataset.}
\format{A matrix with 25,697 gene expression values (columns) and 208 samples (rows).}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE48760}
}
\usage{
GSE48760
}
\description{
Transcriptomics analysis of left ventricles of mouse subjected to an isoproterenol challenge. Last column as variable response.
}
\keyword{datasets}
