% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE5325}
\alias{GSE5325}
\title{GSE5325 dataset.}
\format{A matrix with 27,649 gene expression values (columns) and 106 samples (rows). Last column as variable response.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE5325}
}
\usage{
GSE5325
}
\description{
Analysis of breast cancer tumor samples using 2-color cDNA microarrays.
}
\keyword{datasets}
