% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impute}
\alias{impute}
\title{impute}
\usage{
impute(v, method = "mean")
}
\arguments{
\item{v}{Input numeric vector}

\item{method}{One of 'mean' of 'median'}
}
\value{
A vector with missing values imputed.
}
\description{
impute
}
