% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceProcessing.R
\name{mapAA}
\alias{mapAA}
\title{mapAA}
\usage{
mapAA(seqvec)
}
\arguments{
\item{seqvec}{vector of amino acid (3-letter code)}
}
\description{
This function renames amino acid from 3-letter to 1-letter cod
}

