% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_corr}
\alias{plot_corr}
\title{plot_corr}
\usage{
plot_corr(features, corr.method = "pearson")
}
\arguments{
\item{features}{A numeric matrix as input.}

\item{corr.method}{Specifies the type of correlation to be computed (pearson, spearman or kendall)}
}
\value{
A correlation plot.
}
\description{
Simple function to plot a correlation matrix between input features.
}
