% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_pca}
\alias{plot_pca}
\title{plot_pca}
\usage{
plot_pca(features, class, center = TRUE, scale = TRUE, list.plot = TRUE)
}
\arguments{
\item{features}{A numeric matrix as input.}

\item{class}{Response variable as numeric vector.}

\item{center}{Must the features be centered (default TRUE).}

\item{scale}{Must the features be scaled (default TRUE).}

\item{list.plot}{If TRUE (default), return a list the plots summarizing the PCA process. If FALSE, return a simple plot (PC1 vs. PC2).}
}
\value{
A list of plots.
}
\description{
This function plots the Standard Desviation, Proportion of Variances, Accumulative Variance
for each Principal Component, it returns a set of plots.
}
