% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotPCVariances}
\alias{plotPCVariances}
\title{plotPCVariances}
\usage{
plotPCVariances(dat, na.rm = TRUE, ...)
}
\arguments{
\item{dat}{dataframe: first col: PCs index}

\item{na.rm}{ignore the null values}
}
\description{
This function plots the Standard Desviation, Proportion of Variances and Accumulative Variance
for each Principal Component of the dataset, it returns a set of plots
}

