% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceProcessing.R
\name{reformatSequence}
\alias{reformatSequence}
\title{reformatSequence}
\usage{
reformatSequence(seq)
}
\arguments{
\item{seq}{sequence}
}
\description{
This function reformat the sequence to remove inconsistencies
}

