% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceProcessing.R
\name{removePTM}
\alias{removePTM}
\title{removePTM}
\usage{
removePTM(seq)
}
\arguments{
\item{seq}{sequence}
}
\description{
This function reformat the sequence to remove all PTM labers.
Useful for example to compute AAindex Isoelectric point
}

