% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_functions.R
\name{rfeRF}
\alias{rfeRF}
\title{rfeRF}
\usage{
rfeRF(features, class, number.cv = 10, group.sizes = c(1:10, seq(15, 100,
  5)))
}
\arguments{
\item{features}{A numeric matrix as input.}

\item{class}{Response variable as numeric vector. It will be coerced to factor.}

\item{number.cv}{Number of cross-validation folds (10 default). Used during training phase.}

\item{group.sizes}{A numeric vector of integers corresponding to the number of features that should be retained.}
}
\value{
A list the elements. See \code{\link[caret]{rfe}} for more details.
}
\description{
Recursive feature elimination (RFE) method wrapped with a Random Forest (RF) algorithm for feature importance evaluation.
}
\examples{
output <- rfeRF (features, class, number.cv = 10, group.sizes = seq(1,20,2))


}
