% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TNBC}
\alias{TNBC}
\title{Triple-Negative Breast Cancer (TNBC) dataset.}
\format{A matrix with 15,525 protein expression values (columns) and 44 samples (rows). Last column as variable response.}
\source{
\url{http://www.cell.com/cell-reports/abstract/S2211-1247(15)00341-1}
}
\usage{
TNBC
}
\description{
Label-free deep proteome analysis of 44 (samples and technical replicates) human breast samples.
}
\keyword{datasets}
