% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare.matrix}
\alias{compare.matrix}
\title{compare.matrix}
\usage{
compare.matrix(
  input.matrix,
  output.matrix,
  description = "Percent of removed features:"
)
}
\arguments{
\item{input.matrix}{A numeric matrix as input.}

\item{output.matrix}{A numeric matrix as input.}

\item{description}{Short description summarizing this action.}
}
\value{
A message object.
}
\description{
compare.matrix
}
